C89 final fix up
Also use floats for small values, no need for a double when the value is 0<->1

+++ qtBuildSystem/Mu/touchscreen.cpp
@@ -17,7 +17,7 @@ TouchScreen::~TouchScreen(){
 
 }
 
-double TouchScreen::rangeSwap(double newRange, double oldRange, double value){
+float TouchScreen::rangeSwap(float newRange, float oldRange, float value){
    return value / oldRange * newRange;
 }
